CC-768 lint fix

+++ app/components/course-page/delete-repository-modal.ts
@@ -6,13 +6,24 @@ interface Signature {
   Element: HTMLDivElement;
 
   Args: {
-    repository: RepositoryModel;
+    onClose?: () => void;
+    repository?: RepositoryModel;
   };
 }
 
 export default class DeleteRepositoryModalComponent extends Component<Signature> {
   @action
   deleteRepository() {
+    if (!this.args.repository) {
+      return;
+    }
+
     this.args.repository.deleteRepository({});
+  }
+}
+
+declare module '@glint/environment-ember-loose/registry' {
+  export default interface Registry {
+    DeleteRepositoryModal: typeof DeleteRepositoryModalComponent;
   }
 }
